[clear] リファクタ
